+++ src/core/util/array.h
@@ -3,5 +3,5 @@
 namespace array
 {
 	//Если это будет читать Костя, то это его функция :)
-	float* vectorToArray(std::vector<float> vector, size_t& sizeAr);
+	float* vectorToArray(std::vector<float> vector, int& sizeAr);
 }
